refactor: 댓글방 클래스들을 comment 패키지로 분리
