--- android/app/src/main/java/com/zzang/chongdae/presentation/view/CommentRoomViewModel.kt
@@ -1,3 +1,0 @@
-package com.zzang.chongdae.presentation.view
-
-class CommentRoomViewModel
