+++ android/app/src/main/java/com/zzang/chongdae/presentation/view/comment/CommentRoomViewModel.kt
@@ -1,0 +1,3 @@
+package com.zzang.chongdae.presentation.view.comment
+
+class CommentRoomViewModel
