+++ android/app/src/main/java/com/zzang/chongdae/presentation/view/comment/adapter/CommentRoomViewHolder.kt
@@ -1,10 +1,10 @@
-package com.zzang.chongdae.presentation.view.adapter
+package com.zzang.chongdae.presentation.view.comment.adapter
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import com.zzang.chongdae.databinding.ItemCommentRoomBinding
 import com.zzang.chongdae.domain.model.CommentRoom
-import com.zzang.chongdae.presentation.view.CommentRoomViewModel
+import com.zzang.chongdae.presentation.view.comment.CommentRoomViewModel
 
 sealed class CommentRoomViewHolder(
     view: View,
