+++ android/app/src/main/res/layout/fragment_comment_room.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.view.CommentRoomFragment">
+    tools:context=".presentation.view.comment.CommentRoomFragment">
 
     <TextView
         android:id="@+id/tv_comment_text"
